ch03/myproject_virtualenv/src/django-myproject/requirements/_base.txt
Django==2.2.4
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-imagekit==4.0.2
pilkit==2.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
